+++ src/hal/JsonResponseFormatter.php
@@ -1,0 +1,76 @@
+<?php
+
+namespace tecnocen\roa\hal;
+
+use yii\base\Component;
+use yii\helpers\Json;
+use yii\web\ResponseFormatterInterface;
+
+/**
+ * JsonResponseFormatter formats the given data into a JSON or JSONP response content.
+ *
+ * It is used by [[Response]] to format response data.
+ *
+ * To configure properties like [[encodeOptions]] or [[prettyPrint]], you can configure the `response`
+ * application component like the following:
+ *
+ * ```php
+ * 'response' => [
+ *     // ...
+ *     'formatters' => [
+ *         \yii\web\Response::FORMAT_JSON => [
+ *              'class' => 'yii\web\JsonResponseFormatter',
+ *              'prettyPrint' => YII_DEBUG, // use "pretty" output in debug mode
+ *              // ...
+ *         ],
+ *     ],
+ * ],
+ * ```
+ */
+class JsonResponseFormatter extends Component implements
+    ResponseFormatterInterface
+{
+    /**
+     * @var string the Content-Type header for the response
+     */
+    public $contentType = 'application/hal+json';
+
+    /**
+     * @var string the XML encoding. If not set, it will use the value of [[Response::charset]].
+     */
+    public $encoding = 'UTF-8';
+
+    /**
+     * @var int the encoding options passed to [[Json::encode()]]. For more details please refer to
+     * <http://www.php.net/manual/en/function.json-encode.php>.
+     * Default is `JSON_UNESCAPED_SLASHES | JSON_UNESCAPED_UNICODE`.
+     */
+    public $encodeOptions = 320;
+
+    /**
+     * @var bool whether to format the output in a readable "pretty" format. This can be useful for debugging purpose.
+     * If this is true, `JSON_PRETTY_PRINT` will be added to [[encodeOptions]].
+     * Defaults to `false`.
+     */
+    public $prettyPrint = false;
+
+    /**
+     * Formats the specified response.
+     * @param Response $response the response to be formatted.
+     */
+    public function format($response)
+    {
+        if (stripos($this->contentType, 'charset') === false) {
+            $this->contentType .= '; charset='
+                . ($this->encoding ?: $response->charset);
+        }
+        $response->getHeaders()->set('Content-Type', $this->contentType);
+        if ($response->data !== null) {
+            $options = $this->encodeOptions;
+            if ($this->prettyPrint) {
+                $options |= JSON_PRETTY_PRINT;
+            }
+            $response->content = Json::encode($response->data, $options);
+        }
+    }
+}